requirements.txt
google-generativeai>=0.8.0
google-cloud-aiplatform>=1.38.0
google-auth>=2.17.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
ultralytics>=8.0.0
matplotlib>=3.7.0
seaborn>=0.12.0 